move index to concrete file
