+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ProjectUpgradesPage/UpgradeWizard/steps/FundingCycleStep/LaunchFundingCycleForm.tsx
@@ -5,7 +5,7 @@ import { JuiceSwitch } from 'components/inputs/JuiceSwitch'
 import UnsavedChangesModal from 'components/modals/UnsavedChangesModal'
 import RichButton from 'components/buttons/RichButton'
 import { FundingDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/FundingDrawer'
-import NftDrawer from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer'
+import { NftDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer'
 import { RulesDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/RulesDrawer'
 import { TokenDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/TokenDrawer'
 import { ProjectMetadataContext } from 'contexts/shared/ProjectMetadataContext'
