+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/ReconfigureFundingCycleSettingsPage/ReconfigureFundingCycleForm.tsx
@@ -4,7 +4,7 @@ import UnsavedChangesModal from 'components/modals/UnsavedChangesModal'
 import { MemoFormInput } from 'components/Project/PayProjectForm/MemoFormInput'
 import RichButton, { RichButtonProps } from 'components/buttons/RichButton'
 import { FundingDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/FundingDrawer'
-import NftDrawer from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer'
+import { NftDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer'
 import { RulesDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/RulesDrawer'
 import { TokenDrawer } from 'components/v2v3/shared/FundingCycleConfigurationDrawers/TokenDrawer'
 import { CV_V3 } from 'constants/cv'
